+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 2021-01-12
+
+- BREAKING CHANGE: Format of the dataset has changed
+- Search multiple terms at once, search hashtags and terms
+- Enriched user profile information (some information are only available when logged in)
+- Added minimum and max tweet dates
+- Updated SDK version
+- Custom data
+- Powerful extend output / scraper function
+
 ## 2020-11-25
 
 - Remove the need to provide credentials